Move Codex CLI runtime probes behind boundary
